ssmd-mcp/pyproject.toml
[project]
name = "ssmd-mcp"
version = "0.1.0"
description = "MCP server for querying ssmd market data (parquet/GCS + ssmd-data-ts API)"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "duckdb>=1.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ssmd-mcp = "ssmd_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ssmd_mcp"]
